edit_structure
